Adjust the 'nothing to show' text to match AOSP
When we enable the switch to 'enable for all' a text appears to notify
the user that if they want to enable per-app configurations to flip the
switch off. This is very similar to what you see under the Bluetooth menu
and around Android.

This text should match AOSP, otherwise it looks like shit.

Change-Id: I8f99b5b13b99243296f472675e29e7d071796ae4