Remove expanded desktop category
This category was titled 'expanded desktop options' but really there's
no other option other than the list preference to select statusbar/navbar
or both. Add to the fact that there's no other options in the pipe for this
and it just doesn't make sense to keep around. It looks out of place.

Change-Id: Iad28d6e9d80b046c0fc6a0c74edf69cdfc004ec0